cvo: Allow payloads to retrieve images from an alternate location
Allow a payload to specify an alternate image repository to use instead
of using the location the payload specifies. When manifests are being
pre-processed, the following algorithm is applied:

1. search the manifest for any string that matches the name of an image
   defined in image-references (with leading and trailing word breaks)
2. parse the image location, replacing all but the digest with the new
   payloadRepository
3. replace the string in the manifest with the new location

When the CVO applies the payload, images will be fetched from the
override location rather than the location the payload defined. However,
since we preserve the digest, we ensure that the contents still match.