cvo: Report errors to conditions more consistently
Reporting sync errors to the user via the ClusterVersion status is important
for usability - this will be the first place user interfaces check when
errors occur. Improve the exit paths from the core payload sync path to
capture errors.

Improve the consistency of error behavior by having the Failing condition
contain the error text exactly, and have Progressing summarize the error
in the condition in the context of the action the operator was taking
(reconcile, upgrade). Attempt to preserve the status of the Progressing
and Failing conditions when performing the opposite update so as to
keep users from seeing results that strobe.

Improve the error message returned by the payload update step to give
an administrator / developer a reasonable guess at what the failure is
without exposing too many details. The Failing message can be more
detailed than the Progressing message.

Add two metrics to payload syncing to report the number of payload update
errors and the current number of applied vs pending manifests to apply.

Correct minor nits from the last PR.