cvo: Perform status updates in a single thread
The available updates loop creates data that is merged into the status
during the sync loop. Change how the sync loop writes status to be a
simple live lookup and write with a single conflict retry. Modify
the object in place so that subsequent status writes in the same sync
succeed.

Creating the initial ClusterVersion exits the sync loop to avoid conflicts.

Add a guard that detects the most recent status update and ignores
versions of the object in order to avoid reacting to the intermediate
status updates for Progressing or Failing. This reduces conflict loops
significantly.