start: Resolve the release image to an exact value instead of using tag
If the CVO is started with a tag, it's possible the tag can drift from the
initial value and the cluster would start updating at a random time. Instead,
during startup the CVO attempts to identify the precise pull spec as
Kubernetes reports it and use that as the release version. When the reconciler
runs we will update ourself to that version.

Before fetching the new payload, run a pod directly so as to find the resolved
imageID from the running payload before extracting the payload.