cvo: Validate the CVO prior to acting on it
The CVO needs to perform some minimum sanity checking on incoming input
and communicate to users when it is blocked. The previous mechanism of
rejecting the CV on deserialization is only partial validation and doesn't
cover all scenarios. In the future we may want to have the CVO register
as an admission webhook for its own resource.

Add validation immediately after the CVO loads the object from the cache,
verifying that the object that we see has no errors. If it does, write
an Invalid condition to the status and reset the Progressing condition,
then clear the invalid fields so that the sync loop doesn't act on them.

Simplify initial status setting by having the initial status check
normalize the object status for a set of conditions, including available
updates. This reduces the complexity of the CVO main loop slightly.