docs/dev/operators: Document a dynamic-object approach
This content is descended from openshift/installer@9c8a77dd (docs/dev:
Doc on how to integrate an operator, 2018-09-28,
openshift/installer#377), but handling dynamic configuration is a
generic issue, so the docs should live in the generic CVO repository.
This will make these docs more discoverable.  It will also allow the
installer docs to focus on installer-specific issues for the small
subset of operators that need direct installer bindings.