cvo: Label all applied resources with the current version
Knowing that an object has changed is very useful when debugging
the state of a cluster, and the version also allows us to perform
downward API actions from deployments to match the desired version
out into the outcome status.

Refactor the internal structure of the resource builder so that we
can inject testing of dynamic client actions. The interface may
change in the future, so this is merely to improve the testing of
payload application for now.