sync: Do config syncing in the background
The current sync loop is complex to reason about because a long
running operation (the sync) is interspersed with status updates.
A sufficiently long sync blocks all status updates for the duration
and also prevents a user from cancelling an update quickly (you
must wait until the sync fails or completes to go again).

Instead, define a SyncWorker as a background loop that the sync()
method updates with the desired state each time the cluster version
object changes. The worker encapsulates the complex bits of the
previous sync() handler, including treating initial sync of a version
differently from reconciling, reporting incremental status, and
allowing a user to cancel one sync and replace it with another
(although there are scenarios where we may wish to only allow that
if the update has failed).

The sync() handler invokes SyncWorker.Update() which atomically
updates the status and signals the background loop, and then returns
the current sync status. The sync() handler takes the current status
and updates the ClusterVersion status in a much simpler fashion
than before (we remove all the different variants and instead have
a single syncConfigStatus method).

The new code structure makes it easy for the SyncWorker to cache
the contents of the payload until a new update is requested, which
prevents drift and makes subsequent reconciliation cheaper.

The sync() method also now only performs a single status update
per round which is easier to test and reason about. The previous
code which attempted to use status fields to guess at whether we
were reconciling is now a much clearer for loop in the SyncWorker
Start() method.

For better testability, the act of fetching an update payload and
the act of applying the payload are abstracted behind interfaces
which are more easily mocked, and a new scenario driven test is
created to demonstrate tests that were previously more complicated
to execute.

The existing status tests are refactored to reflect that status is
driven solely by the sync worker's status.

Add an integration test that requires a KUBECONFIG with admin
privileges and will drive scenarios against a real cluster.