cvo: Release our leader lease when we are gracefully terminated
On an upgrade the CVO is always restarted but we don't release our
lease, which causes the new CVO pod to have to wait. We should
release our lease on exit.

Kube 1.14 (pr 71490) contains a new flag on leader lease that allows
the caller to step down gracefully. Because backporting that change
to client-go is complicated, we instead emulate the logic. When that
code is available we can simplify down.

cmd: Refactor how the CVO is started so the integration test is consistent

Move the integration test logic into a new package and reuse startup
logic so that we have a much cleaner start command than before and so
that we are testing what we run in the command. Remove rootOpts and
startOpts and replace them with nested operations.

Add a test that verifies we send leader election events.