refactor: Extract payload related functions for clarity
The cvo package is getting a bit full. In preparation for parallel
payload processing where we want to have a readable graph processor
move payload related functions into their own package. The syncTask
struct is renamed to payload.Task.

Record the filename of the manifest when read and store it on Task
for reference in debugging and parallelization.