payload: Create a task graph that can split a payload into chunks
The payload is viewed as a graph of chunks of tasks that must be
executed in serial (roughly corresponding to the manifests defined
by each operator). Using two verbs, Split -- which ensures that
certain tasks cause a happens-before/happens-after for the entire
graph -- and Parallelize -- which tries to run manifests from
different operators at the same run level (0000_70_*) in parallel,
we roughly halve the linear depth of the payload and ensure that
later operators run in parallel while preserving ordering for
the core operators.

Remove requeing because it can change the order of tasks within
a job, which may cause confusion about potential orderings of
actions in the future.

Future variants might more aggressively parallelize tasks.

Refactor the sync worker to use the task graph with a default
parallelism of 8.