metrics: Report the age of the current payload for alerting
One attack vector for clusters is to deny access to new updates or
to deliberately send an older user. This commit ensures the date the
payload was created is reported back through metrics and up via
telemetry. It also adds a new `completed` type for the cluster_version
metric that reports the last successfully applied cluster version and
the date it was successfully applied the first time (from the new
status history field). Finally, we correct the naming for the update
type to match the name of the status field to `desired`.