status: Generation should be set by what the sync worker is doing
The primary mechanism where by a reconciling controller communicates
to a client is its status, and the client has to wait until the
controller observes its write before assuming the controller is ready.

When writing a client that is triggering an upgrade, we need to:

1. Set desiredUpdate and read resulting generation
2. Wait until status.generation == generation from 1
3. Wait until the update that is marked partial in status history is complete

This fixes the sync worker to be the one who propagates generation, so
that a client doesn't see 2 above when syncing a previous generation.