wip: sync: Perform an initial flood fill of the objects in the payload
During install the normal "safe" path for updates works against us
by serializing a bunch of operations that should never fail, or if
they do fail are likely bugs anyway. Exploit the idempotent create
behavior of Kube-like APIs and attempt to create as much of the
payload as possible in a single create pass, halting if we hit any
errors. We then perform a second pass and wait wait for operators
to report status.

The most common reason to halt the initial pass is when the
kube-apiservers are restarted during configuration, or when we
reach a payload object that depends on an aggregated API which is
unlikely to have succeeded.

There is a tiny amount of unsafety around this optimization. One,
deployments that depend on earlier deployments may fail or go into
backoff because those prerequisites have not healed, which might
actually (unlikely, but possible) delay startup more. Two, this
puts more load on the initial startup and may peg CPU or cause
excessive memory use - however that's more likely to benefit us
than hurt. Finally, when updates are delivered that create objects
earlier in the payload (a new operator is added at runlevel
0000_04_X, before any other operators) AND the previously first
operator adds a new object that depends on that new first operator,
the new object may fail because the first operator hasn't yet
completed. In that case, tests should catch it, and we should fix
the operator if necessary.