sync: Report a specific cluster operator error when not available
Cluster operators block most of upgrade, so we should report the
information the cluster operator has available back to the CV status
object via an UpdateError.

Reduce the amount of debug info printed by cluster operator status
waiting to simplify the logs.

While reviewing CV logs from startup, we run the resync loop more
frequently waiting for the CO to be created because we exit the
poll loop on any error. When we see "CO not found", continue the
loop to avoid going back and syncing on top.

Finally, remove the apiVersion info from the status log message,
it didn't make much sense and the end user is going to have to look
at the payload anyway if we get into a wierd error case where it
matters (the apiVersion was put in in case we had to debug a typo
in a version, which really shouldn't happen).