test: Fix flake when retrieving payloads
When a payload is retrieved the version may be empty, and if the
status is updated it'll have an empty version. This caused an
integration flake.

Since retrieving the payload can take a certain amount of time,
reflect that in the progressing message and give a unique reason.
The test should then verify the state.