operators: Retry connection refused errors while waiting for CO
During install and update the kube-apiserver may be updated one or
more times, and since the CVO is configured to talk to localhost the
server may be unavailable. If we see connection refused, treat it
as retriable at this level vs propagating it up, since we can't
talk to another server.

A more long term fix is for the rest client to accept multiple
possible servers and try all of them, which requires significant
changes to client-go.