sync: Distinguish between reconcile, initial, and sync states
Identify what sort of change we are making to the cluster in order
to better handle the different states. Does not change status reporting
or the internal sync which will be in a follow up commit.

The three states are:

* Initializing - we haven't yet completed a payload
* Updating - we're moving between two payloads of different versions
* Reconciling - we're level on a completed payload

The states will allow us to more efficiently reconcile, start faster,
and be safer during upgrades.