sync: Change from retry up to X times to continuous retry until cancel
The sync loop currently mixes a combination of:

1. Try forever (top level sync loop)
2. Cancel sync when a timeout is reached
3. Retry individual Apply actions in the sync workers up to a max retry
4. Retry within Apply actions

This led to bugs in initialization where transient errors cause a
prereq step to fail and never complete, leading to other operators
never going live.

This commit simplifies the structure of our retry logic to be:

1. Try to sync forever
2. Cancel sync after a certain period of failure and go into backoff
3. Retry individual Applies forever with capped backoff
4. Retry applies as long as context is not cancelled

This implies all top level sync has a context with a deadline which
is the case today - some test cases are changed for it.

This should also reduce the lag between cancel (on user action, for
instance) and the sync loop terminating, since everything is context
gated.