Don't create new version histories for empty version
If images match, the old version is set, and the new version is
unset we should not create a new history entry. The scenario is
during an upgrade or other rescheduling of the apiserver pod where
we don't have the version pulled and the input was an image, so
we briefly regress to a "have image but no version" state until
the new controller instance can retrieve the payload.

Add test and correctly handle the scenario. Leave some debugging
code in case this crops up in the future.