metrics: Report when the CVO is transitioning a payload
The current metric for 'desired' cluster_version reports the user's
input. However this metric doesn't discriminate between an invalid
update (where nothing will be happening) and when the operator is
attempting to converge, nor does it report the version number we
read from the release image.

Add a new 'updating' cluster_version type that is returned if and
only if the most recent history entry is in the Partial update state
(i.e., not reconciling), which covers both upgrade and rollback. The
image and version are reported from the history and should be
accurate to the payload's internal version. The value is the start
time in unix seconds of this current version and is a good proxy for
"how long have we been trying to upgrade".

Identified by the analysis team because 'desired' is imprecise.