pkg/cincinnati: Soften unknown-version message
Maybe having the current version not in the desired channel's graph is
a problem, but maybe you know what you're doing.  And Cincinnati might
*know* about your version, and just consider it to be outside your
channel, in which case the solution may be "switch channels" and not
"switch versions".  This change hopefully makes it a bit easier for an
admin in this condition to figure out the appropriate next step for
their cluster.