sync: During reconcile, try different parts of the payload and don't hang
The worst outcome of reconcile would be that a break occurs in a late
component and the CVO gets hung trying to reach it. Instead, when reconciling
take the graph and run it in random blocks at a lower parallelization
factor to ensure that we will eventually try all blocks, even if something
fundamental is broken. The permutation of the graph ordering is stable
for a given payload and sync time (so that when we test locally we are
testing the same order a customer we would see).

The only gap still left would be a component breaks that can't be recovered
due to ordering (to fix a role binding we need to deploy a deployment, but
the lack of the deployment prevents us from sending the role binding) which
we can potentially investigate later by randomizing the task order.

We track "number of attempts" in a given state / payload to see the input.