verifier: Add public key verification of release image digests
Use the existing Atomic image signing protocol to read detached signatures
for images by digest from a remote location so that release images can
be verified before they are executed.

Add a Verifier interface to the CVO that abstracts checking for verified
updates. On start, check the payload for a config map with the annotation
release.openshift.io/verification-config-map set (value is ignored) and
load the set of all public keys that must be verified along with the http
or file store locations for detached signatures. Every key must be
verified to accept the payload. A subsequent commit will leverage the
Verifier to block downloading a new release image.