sync: Before retrieving a payload, verify it and report errors
The verifier is loaded at startup and passed to the payload retriever,
which then delegates to the verifier and returns an error if the
digest is not authentic. If no verifier is configured, the default
verifier is to reject the payload as inauthentic.

Move the initialization of the sync worker to after the verifier is
loaded so that we can pass the verifier into the payload retriever,
which makes the New() method slightly smaller and separates out
payload related actions from controller related actions.