install/0000_00_cluster-version-operator_01_clusteroperator.crd: Degraded -> Failing
Fix [1]:

  $ oc get clusteroperators kube-apiserver
  NAME             VERSION                             AVAILABLE   PROGRESSING   FAILING   SINCE
  kube-apiserver   4.1.0-0.nightly-2019-04-22-005054   True        False                   20m

when the ClusterOperator only sets Degraded.  This should have
happened in 94c4e576d1 (Use the new degraded condition is available,
2019-04-16, #165).  There may be some JSONPath syntax to collapse both
into a single column:

  .status.conditions[?(@.type IN ["Degraded", "Failing"])].status

but if it exists, I haven't discovered it.  We're hopefully going to
fix the operators that are still using Failing soon, so just skip to
the new field without supporting both [2].

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1702087
[2]: https://github.com/openshift/cluster-version-operator/pull/172#issuecomment-485603684