status: Simplify operator not available status
In practice, the output of displaying version operands is redundant
and our current use doesn't require it. Since the core output is

"we are going to X, these are the things not there yet"

also showing

"we are going to X, these are the things not there yet A-X,B-X,C-X"

makes it harder to understand what is going on.

Strip all sub version reporting from cluster operators, and in the
future we can improve this if we have a use case.