status: Hide generic operator status in favor of more specific errors
When upgrading and a sync loop exits, we often see:

* Operator A not upgraded yet
* Operator B not upgraded yet
* Operator C not upgraded yet
* Some more specific error
* Operator D not upgraded yet

Since the not available messages carry little value and are expected
when we are processing parallel output, if we have one or more errors
during sync that are not the generic error, only report those. This
focuses the user's attention on the actual problem, rather than
distracting them with noise. In the ideal case, the user immediately
sees "router is totally broken" vs "a, b, c, d, e, router is broken, f".