docs/user/reconciliation: Document release-image application
To make it easier for admins and support staff to troubleshoot when
the cluster-version operator gets stuck.

The SVGs were generated with:

  $ go build ./hack/cluster-version-util
  $ mkdir /tmp/release
  $ oc image extract quay.io/openshift-release-dev/ocp-release:4.1.0[-1] --path /:/tmp/release
  $ mkdir /tmp/release/manifests
  $ ./cluster-version-util task-graph /tmp/release | dot -Tsvg >docs/user/tasks-by-number-and-component.svg
  $ ./cluster-version-util task-graph --parallel flatten-by-number-and-component /tmp/release | dot -Tsvg >docs/user/tasks-flatten-by-number-and-component.svg

using:

  $ dot -V
  dot - graphviz version 2.30.1 (20170916.1124)

I initially put the utility program in cmd/cluster-version-util,
moving the main operator to cmd/cluster-version-operator.  But Abhinav
was concerned about implied support for tooling that is really just
intended for exposing internal CVO logic, so this commit puts it under
hack/ with a caveat in the main help text.

The PATH documentation is stuffed into Use because cobra lacks direct
support for named positional arguments [1].

[1]: https://github.com/spf13/cobra/issues/378