metrics: Add cluster_version{type="initial"}, change "cluster"
An important query is "how old are the clusters of a specific version"
and the current "cluster" query does not make this possible. In addition,
we can't get initial + current version in a single query. Split the
use case for "cluster" into "initial" and "cluster", where "initial"
is the same as "cluster" today, and "cluster" now points to the same
image and version as "current", has from_version populated, and has
the same value.