vendor: remove constraint from glog
Back in 6c1716f, we dropped the direct dependency on glog, which
resulted in `dep ensure` showing the following warning:

    Warning: the following project(s) have [[constraint]] stanzas in
    Gopkg.toml:

      ✗  github.com/golang/glog

It also resulted in the constraint no longer being respected
(`[[override]]` is required for transitive dependencies). Since we have
been using a newer version of glog for almost two months now, I'm
dropping the constraint instead of converting into an override.