payload/render: update the bootstrap render to skip servicemonitor
the `manifests` directory on the bootstrap is used by the cluster-bootstrap to push to the cluster.
`servicemonitor` for cvo was added by https://github.com/openshift/cluster-version-operator/pull/214
`servicemonitor` api is created by the cluster-monitoring-operator and therefore this causes the bootstrapping to get stuck until we get the monitoring operator running.

This skips the `servicemonitor` in the bootstrap render as it is not required for the bootstrap cvo pod.