cincinnati: add arch parameter to requests
OpenShift is going to be gaining support for architectures other than
AMD64 soon. In order to make sure that the updates provided to a cluster
are compatible with the underlying architecture(s), we are adding an
"arch" parameter to the Cincinnati upstream queries. This new parameter
will contain a set of architectures, separated by commas, that need to
be supported by a release image in order for it to be applied to the
cluster. In a homogeneous deployment, the set of architectures will only
contain a single architecture (this is what this patch implements).
Eventually, once heterogeneous deployments are supported, the set may
contain multiple architectures. The details of heterogeneous deployments
and their constraints haven't been fully fleshed out, so this scheme may
change over time.