pkg/cvo/metrics: Report ClusterVersion conditions with reasons
This will allow us to discover upgrade and other failure reasons
without having to resort to a must-gather or similar [1].  And also to
look at any other version conditions in Telemetry.

Stick this in cluster_operator_conditions, since we already have a
'reason' slot there.  And ClusterVersion.Status.Conditions is pretty
much the same thing as ClusterOperator.Status.Conditions; we'll want
to see all of those.  I don't see a reason to add a new metric to
separate cluster-version operator failures from second-level operator
failures; the name should be sufficient for that.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1741645