cmd: Add --telemetry-id option
As an alternative to:

* removing the ClusterVersion defaulting [1], or
* moving the telemetry ID out into a Telemetry-specific config object [2]

this commit allows us to address the race between the
installer-created ClusterVersion and the
cluster-version-operator-generated default ClusterVersion by removing
the former while still allowing an installer-generated Telemetry ID.

We currently need an installer-generated Telemetry ID to support
things like OpenShift Dedicated (OSD) [3], where we currently use the
telemetry ID that the installer exposes in metadata.json [4] to feed
the cluster manager [5].  There is concern that if Hive [6] extracted
the Telemetry ID from the running cluster (vs. feeding in an
installer-chosen ID), that this would race with the cluster's reported
Telemetry and that, for example, Hive-launched OSD Telemetry might be
erroneously claimed to be a user-provisioned cluster.

I'm less clear on why we don't prefer pushing the Telemetry ID into a
Telemetry-specific config.  It seems like it would be less work than
this commit (which has a lot of middleman plumbing).  I'm not entirely
clear on the API we're committed to supporting for ClusterVersion's
spec.clusterID property.  But this CVO plumbing was what we were able
to form a good-enough-for-now consensus around, and we can come back
later with a Telemetery-config approach if we can build consensus
around that.

[1]: https://github.com/openshift/cluster-version-operator/pull/238
[2]: https://github.com/openshift/cluster-version-operator/pull/238#issuecomment-522350796
[3]: https://www.openshift.com/products/dedicated/
[4]: https://github.com/openshift/installer/blob/4e204c5e509de1bd31113b0c0e73af1a35e52c0a/pkg/types/clustermetadata.go#L17-L18
[5]: https://cloud.redhat.com/openshift/
[6]: https://github.com/openshift/hive/