pkg/payload/image: avoid rendering manifests
The `image` command currently loads the image-references file and then
renders all manifests. Clearly the manifest rendering can be avoided
when we just want to look up an image reference.

This command is currently run 17 times by bootkube, and my quick test
shows this refactor halving the roughly 10 seconds this takes.

The main refactoring here is to split the loadUpdatePayloadMetadata()
function into getPayloadTasks() and loadImageReferences() so that the
latter can be used standalone in this particular case.